ADD: Add GinIndex to entry author field

+++ talkwalker/models.py
@@ -12,12 +12,29 @@ import json
 from django.contrib.postgres.indexes import GinIndex, OpClass
 from django.db.models.functions import Upper
 
+import environ
+
+env = environ.Env()
+ALLOWED_HOSTS = [env('ALLOWED_HOSTS')]
+
 
 class TalkwalkerFeedlink(models.Model):
     country = models.CharField('Country', max_length=200, null=True, blank=True)
     source1 = models.CharField('Source1', max_length=200, null=True, blank=True)
     sourceurl = models.URLField(max_length=200, null=True, blank=True)
     alexaglobalrank = models.BigIntegerField(default=0)
+    
+    class Meta:
+        indexes = [
+            GinIndex(
+                OpClass(Upper("country"), name="gin_trgm_ops"),
+                name="tlw_country_gin_index",
+            ),
+            GinIndex(
+                OpClass(Upper("source1"), name="gin_trgm_ops"),
+                name="tlw_source1_gin_index",
+            )
+        ]
 
 
 class TalkwalkerPost(models.Model):
@@ -57,66 +74,76 @@ class TalkwalkerPost(models.Model):
                 OpClass(Upper("entry_summary"), name="gin_trgm_ops"),
                 name="tlw_entry_summary_gin_index",
             ),
+            GinIndex(
+                OpClass(Upper("entry_author"), name="gin_trgm_ops"),
+                name="tlw_entry_author_gin_index",
+            )
         ]
 
 
 @receiver(post_save, sender='twenty_four_seven.ProjectTwentyFourSeven')
 @receiver(post_save, sender=Project)
 def create_periodic_task(sender, instance, created, **kwargs):
-    if created:
-        Livestream(instance.id, sender.__name__).create()
-        crontab_schedule = CrontabSchedule.objects.create(
-            minute='*/20',
-            hour='*',
-            day_of_week='*',
-            day_of_month='*',
-        )
-        PeriodicTask.objects.create(
-            crontab=crontab_schedule,
-            name=f'LiveSearch_project_{instance.id}',
-            task='talkwalker.tasks.livesearch_sender',
-            args=json.dumps([instance.id, sender.__name__]),
-        )
+    if ALLOWED_HOSTS[0] != 'localhost':
+        if created:
+            Livestream(instance.id, sender.__name__).create()
+            crontab_schedule = CrontabSchedule.objects.create(
+                minute='*/20',
+                hour='*',
+                day_of_week='*',
+                day_of_month='*',
+            )
+            PeriodicTask.objects.create(
+                crontab=crontab_schedule,
+                name=f'LiveSearch_project_{instance.id}',
+                task='talkwalker.tasks.livesearch_sender',
+                args=json.dumps([instance.id, sender.__name__]),
+            )
 
 
 @receiver(post_save, sender='twenty_four_seven.ProjectTwentyFourSeven')
 def tfs_items(sender, instance, created, **kwargs):
-    if created:
-        crontab_schedule = CrontabSchedule.objects.create(
-            minute='*/10',
-            hour='*',
-            day_of_week='*',
-            day_of_month='*',
-        )
-        PeriodicTask.objects.create(
-            crontab=crontab_schedule,
-            name=f'Attach_items_tfs_{instance.id}',
-            task='twenty_four_seven.models.attach_online_posts',
-            args=json.dumps([instance.id]),
-        )
+    if ALLOWED_HOSTS[0] != 'localhost':
+        if created:
+            crontab_schedule = CrontabSchedule.objects.create(
+                minute='*/10',
+                hour='*',
+                day_of_week='*',
+                day_of_month='*',
+            )
+            PeriodicTask.objects.create(
+                crontab=crontab_schedule,
+                name=f'Attach_items_tfs_{instance.id}',
+                task='twenty_four_seven.models.attach_online_posts',
+                args=json.dumps([instance.id]),
+            )
 
 
 @receiver(post_save, sender='twenty_four_seven.ProjectTwentyFourSeven')
 @receiver(post_save, sender=Project)
 def fetch_talkwalker_posts(sender, instance, created, **kwargs):
-    if created:
-        Asker(instance.id, sender.__name__).run()
+    if ALLOWED_HOSTS[0] != 'localhost':
+        if created:
+            Asker(instance.id, sender.__name__).run()
 
 
 @receiver(pre_delete, sender=Project)
 @receiver(pre_delete, sender='twenty_four_seven.ProjectTwentyFourSeven')
 def delete_livestream(sender, instance, **kwargs):
-    Livestream(instance.id, sender.__name__).delete()
+    if ALLOWED_HOSTS[0] != 'localhost':
+        Livestream(instance.id, sender.__name__).delete()
 
 
 @receiver(pre_delete, sender='twenty_four_seven.ProjectTwentyFourSeven')
 def delete_periodic_tasks(sender, instance, **kwargs):
-    tasks = PeriodicTask.objects.filter(name=f'Attach_items_tfs_{instance.id}')
-    tasks.delete()
+    if ALLOWED_HOSTS[0] != 'localhost':
+        tasks = PeriodicTask.objects.filter(name=f'Attach_items_tfs_{instance.id}')
+        tasks.delete()
 
 
 @receiver(pre_delete, sender=Project)
 @receiver(pre_delete, sender='twenty_four_seven.ProjectTwentyFourSeven')
 def delete_live_search_periodic_tasks(sender, instance, **kwargs):
-    tasks = PeriodicTask.objects.filter(name=f'LiveSearch_project_{instance.id}')
-    tasks.delete()
+    if ALLOWED_HOSTS[0] != 'localhost':
+        tasks = PeriodicTask.objects.filter(name=f'LiveSearch_project_{instance.id}')
+        tasks.delete()
